Transports: deck map ids, boarding by teleport, and the crew a deck death used to kill
Ported from mangos_zero f5b943a2 (".tele command fixes"), checked hunk by
hunk against this tree rather than applied.

THE CRASH. A vessel deck map carries a minted id well past 65535, and the
map id was truncated to uint16 in four places on the way in:
MapPersistentState and its three subclasses, CreatureData::mapid,
GameObjectData::mapid, and the ObjectMgr cell-spawn accessors. The
persistent-state constructor is the fatal one: GetMapEntry() looked up
1020808 & 0xFFFF = 37768, found nothing, and every caller that dereferenced
the result crashed. Killing anything on a deck was enough.

Two divergences from the source commit, both because this tree differs:

 - the per-cell spawn store is keyed by MAKE_PAIR32(mapid, spawnMode), and
   that macro casts the low half to uint16 -- widening the parameter alone
   would have moved the truncation into the key. The key is now a uint64
   built by SpawnCellKey(), so map 1020808 and map 37768 no longer share
   one cell store. mangos_zero has no spawnMode and no such key.
 - HandleGoHelper takes z and orientation as POINTERS here, so it has no
   non-positive sentinel to misread on a hull; a deck only needs the world
   fallbacks skipped, and an explicit z demanded.

Boarding from anywhere: TransportMap::Board writes the deck spot into the
passenger movement state and ports his client to the water she sails, which
is the login path exactly. Player::TeleportTo routes a vessel map id there --
naming one on the wire kills the client in CMap::LoadWdt() -- and keeps
MOVEFLAG_ONTRANSPORT when the port is one that carries its passengers.

Also: the vessel maps are minted before the spawn tables are read, or every
crew row is dropped as pointing at a map sMapStore does not know; login
sends the deck offset rather than the hull world pose; a summon on a deck
takes its spot from the hull mesh instead of the sea floor a hundred yards
below; Creature::CleanupsBeforeDelete unboards a dying pet before the crew
index is left holding it; and .gps reports the deck half.

+++ src/game/Object/GameObject.h
@@ -613,7 +613,7 @@ enum GOState
 struct GameObjectData
 {
     uint32 id;                                              // entry in gameobject_template
-    uint16 mapid;
+    uint32 mapid;                                           // uint32 for the same reason as CreatureData::mapid
     uint16 phaseMask;
     float posX;
     float posY;
